pyproject.toml
[tool.poetry]
name = "python-libs"
version = "0.1.0"
description = ""
authors = ["seungjinhan <hanblues@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest-mock = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
